Move files around and add main command
I realized the files were in the wrong place.

Also this adds a main command to run the CRS processing.
